# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: Mini Banking Platform API
  description: |
    A simplified banking platform API that handles financial transactions with
    double-entry bookkeeping. The system maintains data integrity through ledger
    validation and provides user-friendly error responses following RFC 7807.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
...
paths:
...
  /transactions/exchange:
    post:
      tags:
        - Transactions
      summary: Exchange currency within user's accounts
      description: |
        Exchanges money between the authenticated user's accounts with different currencies.
        Uses a fixed exchange rate: 1 USD = 0.92 EUR. The operation is atomic and maintains
        double-entry bookkeeping integrity through cashbook accounts.
      operationId: exchange
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRequest'
      responses:
        '200':
          description: Exchange completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeResponse'
        '400':
          description: Invalid exchange request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                negativeAmount:
                  summary: Negative amount
                  value:
                    type: "https://minibankingplatform.com/problems/negative-exchange-amount"
                    title: "Negative Exchange Amount"
                    status: 400
                    detail: "Cannot exchange a negative amount: -100.00"
                    instance: "/transactions/exchange"
                    amount: "-100.00"
                sameCurrency:
                  summary: Same currency exchange
                  value:
                    type: "https://minibankingplatform.com/problems/same-currency-exchange"
                    title: "Same Currency Exchange"
                    status: 400
                    detail: "Cannot exchange within the same currency: USD"
                    instance: "/transactions/exchange"
                    currency: "USD"
                insufficientFunds:
                  summary: Insufficient funds
                  value:
                    type: "https://minibankingplatform.com/problems/insufficient-funds"
                    title: "Insufficient Funds"
                    status: 400
                    detail: "Account has insufficient funds for this exchange"
                    instance: "/transactions/exchange"
                    available: "500.00"
                    required: "1000.00"
                    currency: "USD"
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Account not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoints
  schemas:
...
    ExchangeRequest:
      type: object
      required:
        - sourceAccountId
        - targetAccountId
        - amount
      properties:
        sourceAccountId:
          type: string
          format: uuid
          description: Source account UUID (currency to exchange from)
          x-oapi-codegen-extra-tags:
            validate: "required,uuid"
        targetAccountId:
          type: string
          format: uuid
          description: Target account UUID (currency to exchange to)
          x-oapi-codegen-extra-tags:
            validate: "required,uuid"
        amount:
          type: string
          pattern: ^\d+(\.\d{1,2})?$
          description: Amount to exchange from source currency
          example: "100.00"
          x-oapi-codegen-extra-tags:
            validate: "required"

    # Response schemas
...
    Money:
      type: object
      properties:
        amount:
          type: string
          description: Decimal amount with 2 decimal places
          example: "1000.00"
        currency:
          $ref: '#/components/schemas/Currency'
...
    ExchangeResponse:
      type: object
      properties:
        transactionId:
          type: string
          format: uuid
        sourceAccountId:
          type: string
          format: uuid
        targetAccountId:
          type: string
          format: uuid
        sourceAmount:
          $ref: '#/components/schemas/Money'
        targetAmount:
          $ref: '#/components/schemas/Money'
        exchangeRate:
          type: string
          description: Exchange rate used (e.g., "0.92" for USD to EUR)
          example: "0.92"
        timestamp:
          type: string
          format: date-time
...
    Currency:
      type: string
      enum:
        - USD
        - EUR
      description: Supported currencies
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: |
            A URI reference that identifies the problem type.
            This specification encourages that, when dereferenced,
            it provide human-readable documentation for the problem type.
          example: "https://minibankingplatform.com/problems/insufficient-funds"
        title:
          type: string
          description: |
            A short, human-readable summary of the problem type.
            It SHOULD NOT change from occurrence to occurrence of the problem.
          example: "Insufficient Funds"
        status:
          type: integer
          description: |
            The HTTP status code generated by the origin server for this
            occurrence of the problem.
          example: 400
        detail:
          type: string
          description: |
            A human-readable explanation specific to this occurrence
            of the problem.
          example: "Account has insufficient funds for this transfer"
        instance:
          type: string
          format: uri-reference
          description: |
            A URI reference that identifies the specific occurrence of
            the problem. It may or may not yield further information
            if dereferenced.
          example: "/transactions/transfer"
      additionalProperties: true
      example:
        type: "https://minibankingplatform.com/problems/insufficient-funds"
        title: "Insufficient Funds"
        status: 400
        detail: "Account has insufficient funds for this transfer"
        instance: "/transactions/transfer"
        available: "500.00"
        required: "1000.00"
        currency: "USD"